# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Workforce Analytics & IT Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/summary:
    get:
      operationId: getSummary
      tags: [reports]
      summary: Dashboard overview KPIs
      parameters:
        - name: group
          in: query
          required: false
          description: Restrict aggregation to devices in this group
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            description: Range start day in YYYY-MM-DD format (inclusive); defaults to today
        - name: to
          in: query
          required: false
          schema:
            type: string
            description: Range end day in YYYY-MM-DD format (inclusive); defaults to today
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
  /reports/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [reports]
      summary: Per-device productivity over a date range
      parameters:
        - name: group
          in: query
          required: false
          description: Restrict aggregation to devices in this group
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            description: Range start day in YYYY-MM-DD format (inclusive); defaults to today
        - name: to
          in: query
          required: false
          schema:
            type: string
            description: Range end day in YYYY-MM-DD format (inclusive); defaults to today
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardItem"
  /reports/group-comparison:
    get:
      operationId: getGroupComparison
      tags: [reports]
      summary: Per-group productivity comparison over a date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            description: Range start day in YYYY-MM-DD format (inclusive); defaults to today
        - name: to
          in: query
          required: false
          schema:
            type: string
            description: Range end day in YYYY-MM-DD format (inclusive); defaults to today
      responses:
        "200":
          description: Group comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupComparisonItem"
...
components:
...
  schemas:
...
    SummaryResponse:
      type: object
      properties:
        devices:
          type: object
          properties:
            total:
              type: integer
            online:
              type: integer
          required: [total, online]
        usersCount:
          type: integer
        screenshotsToday:
          type: integer
        pendingCommands:
          type: integer
        activityToday:
          type: object
          properties:
            productiveSeconds:
              type: integer
            unproductiveSeconds:
              type: integer
            neutralSeconds:
              type: integer
            undefinedSeconds:
              type: integer
            totalSeconds:
              type: integer
          required:
            - productiveSeconds
            - unproductiveSeconds
            - neutralSeconds
            - undefinedSeconds
            - totalSeconds
      required:
        - devices
        - usersCount
        - screenshotsToday
        - pendingCommands
        - activityToday
    LeaderboardItem:
      type: object
      properties:
        deviceId:
          type: string
          format: uuid
        systemName:
          type: string
        productiveSeconds:
          type: integer
        totalSeconds:
          type: integer
        score:
          type: integer
      required: [deviceId, systemName, productiveSeconds, totalSeconds, score]
    GroupComparisonItem:
      type: object
      properties:
        group:
          type: string
        deviceCount:
          type: integer
        productiveSeconds:
          type: integer
        totalSeconds:
          type: integer
        score:
          type: integer
      required: [group, deviceCount, productiveSeconds, totalSeconds, score]
...